requirements.txt
django
djangorestframework
rest_framework_simplejwt
djangorestframework-simplejwt
pillow